// Client/src/app/features/assets/data-access/models/create-asset-request.ts
import { AssetStatus } from "./asset-status";
import { AssetType } from "./asset-type";

export interface CreateAssetRequest {
  readonly assetName: string;
  readonly assetType: AssetType;
  readonly assetStatus: AssetStatus;
  readonly manufacturer: string;
  readonly model: string;
  readonly serialNumber: string;
  readonly macAddress: string | null;
  readonly wifiMacAddress: string | null;
  readonly imei: string | null;
  readonly operatingSystem: string | null;
  readonly operatingSystemVersion: string | null;
  readonly assignedStoreId: string;
  readonly assignedUserId: string | null;
  readonly rfidTagId: string | null;
}